Add files via upload
Create a form with 2 labels, 2 textboxes (type="text", not type="number"), and a button. The labels should ask the user to enter a number. When the user clicks the button, first verify that the user entered numbers in both textboxes. If they entered numbers, divide the first number by the second number and display the result to the nearest hundredth. Also display a message if the second number is 0.